messages: split monolithic e2e test file into focused modules
Split the 3900-line messages_e2e.rs into 7 focused test files
organized by test category: convergence, restart, multi_conversation,
giftwrap, relay_list, multi_account, and connection. Moved shared
relay-level polling helpers into harness/relay.rs.

Changelog-None:
